# Conflicts:
#	src/config/ajax-path.js

+++ src/component/hsieh/Eachrecipe/Eachrecipe.js
@@ -14,14 +14,34 @@ function Eachrecipe() {
         recipes_type: '',
         recipes_cooking_degree: '',
         recipes_ingredient: '',
-        recipes_cooking_method: '',
+        recipes_ingredient1: '',
+        recipes_ingredient2: '',
+        recipes_ingredient3: '',
+        recipes_ingredient4: '',
+        recipes_ingredient5: '',
+        recipes_ingredient6: '',
+        recipes_ingredient7: '',
+        recipes_ingredient8: '',
+        recipes_ingredient9: '',
+        recipes_step: '',
+        recipes_step1: '',
+        recipes_step2: '',
+        recipes_step3: '',
+        recipes_step4: '',
+        recipes_step5: '',
+        recipes_step6: '',
+        recipes_step7: '',
+        recipes_step8: '',
+        recipes_step9: '',
         recipes_description: '',
         recipes_img: '',
-        cooking_create_member_Id: '',
+        recipe_creater: '',
         recipes_collection: '',
         recipes_like: '',
         created_at: '',
     });
+
+    const loginUser = JSON.parse(localStorage.getItem('auth'));
 
     const navigate = useNavigate();
 
@@ -45,22 +65,27 @@ function Eachrecipe() {
 
     // }, []);  useEffect基本架構
 
-    const Delectrecipe = async () => {
+    const Deleterecipe = async () => {
         const data = {
             recipes_sid: params.recipes_sid,
         };
-        const r = await fetch('http://localhost:3600/recipe/delete', {
-            method: 'DELETE',
-            body: JSON.stringify(data),
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        });
-        // console.log('abc:');
-        const obj = await r.json();
-        console.log(obj);
-        alreadydelete(obj);
-        console.log(alreadydelete);
+
+        if (loginUser.customer_id != eachrecipe.customer_id) {
+            alert('您並非本食譜提供者');
+            navigate('/recipe', { replace: true });
+        } else {
+            const r = await fetch('http://localhost:3600/recipe/delete', {
+                method: 'DELETE',
+                body: JSON.stringify(data),
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            });
+            const obj = await r.json();
+            console.log(obj);
+            alreadydelete(obj);
+            console.log(alreadydelete);
+        }
     };
 
     function alreadydelete(obj) {
@@ -71,6 +96,25 @@ function Eachrecipe() {
             navigate('/recipe', { replace: true });
         }
     }
+
+    function sendtoupdate() {
+        if (loginUser.customer_id != eachrecipe.customer_id) {
+            alert('您並非本食譜提供者');
+            navigate('/recipe', { replace: true });
+        } else {
+            navigate(`/recipe/updaterecipe/${params.recipes_sid}`, {
+                replace: false,
+            });
+        }
+    }
+
+    // function like() {
+
+    // }
+
+    // function collection() {
+
+    // }
 
     return (
         <>
@@ -85,7 +129,7 @@ function Eachrecipe() {
                 <div className="recipedetailineach">
                     <div className="autherineach">
                         <p className="authernameineach">
-                            作者：{eachrecipe.cooking_create_member_Id}
+                            作者：{eachrecipe.recipe_creater}
                         </p>
 
                         <div className="likeandcollectineach">
@@ -112,7 +156,7 @@ function Eachrecipe() {
                     <div className="recipeinfomationineach">
                         <div className="recipedataineach">
                             <div className="greencircleineach">
-                                <img src="/images/clock.svg" alt="" />
+                                <img src="/images/clock1.svg" alt="" />
                             </div>
                             <div>
                                 <p>
@@ -128,7 +172,7 @@ function Eachrecipe() {
 
                         <div className="recipedataineach">
                             <div className="greencircleineach">
-                                <img src="/images/heat.svg" alt="" />
+                                <img src="/images/heat1.svg" alt="" />
                             </div>
                             <div>
                                 <p>
@@ -199,9 +243,140 @@ function Eachrecipe() {
                     <div className="materialineach">
                         <div className="blackineachrecipe">使用食材</div>
                         <div className="showarea1ineach">
-                            <div>
-                                <p className="chapter1ineach">
-                                    {eachrecipe.recipes_ingredient}
+                            <p className="chapter1ineach">
+                                {eachrecipe.recipes_ingredient}
+                                {/* 食材1 */}
+                            </p>
+                            {/* 分隔線 */}
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient1.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient1}
+                                    {/* 食材2 */}
+                                </p>
+                            </div>
+
+                            {/* 分隔線 */}
+
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient2.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient2}
+                                    {/* 食材3 */}
+                                </p>
+                            </div>
+
+                            {/* 分隔線 */}
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient3.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient3}
+                                    {/* 食材4 */}
+                                </p>
+                            </div>
+
+                            {/* 分隔線 */}
+
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient4.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient4}
+                                    {/* 食材5 */}
+                                </p>
+                            </div>
+
+                            {/* 分隔線 */}
+
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient5.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient5}
+                                    {/* 食材6 */}
+                                </p>
+                            </div>
+
+                            {/* 分隔線 */}
+
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient6.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient6}
+                                    {/* 食材7 */}
+                                </p>
+                            </div>
+
+                            {/* 分隔線 */}
+
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient7.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient7}
+                                    {/* 食材8 */}
+                                </p>
+                            </div>
+
+                            {/* 分隔線 */}
+
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient8.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient8}
+                                    {/* 食材9 */}
+                                </p>
+                            </div>
+
+                            {/* 分隔線 */}
+
+                            <div
+                                className={
+                                    eachrecipe.recipes_ingredient9.length === 0
+                                        ? 'chapter1ineach1'
+                                        : 'chapter1ineach'
+                                }
+                            >
+                                <p>
+                                    {eachrecipe.recipes_ingredient9}
+                                    {/* 食材10 */}
                                 </p>
                             </div>
                         </div>
@@ -212,12 +387,151 @@ function Eachrecipe() {
                     <div className="cookingineach">
                         <div className="greenineachrecipe">料理方式</div>
                         <div className="showarea2ineach">
+                            {/* 步驟呈現的區域 */}
                             <div className="numberineachrecipe">1</div>
                             <div className="chapter2ineach">
-                                {eachrecipe.recipes_cooking_method}
+                                {eachrecipe.recipes_step}
+                                {/* 步驟1 */}
                             </div>
                         </div>
-                        {/* 引入料理方式 */}
+
+                        {/* 分隔線 */}
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step1.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">2</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step1}
+                                {/* 步驟2 */}
+                            </div>
+                        </div>
+
+                        {/* 分隔線 */}
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step2.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">3</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step2}
+                                {/* 步驟3 */}
+                            </div>
+                        </div>
+
+                        {/* 分隔線 */}
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step3.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">4</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step3}
+                                {/* 步驟4 */}
+                            </div>
+                        </div>
+
+                        {/* 分隔線 */}
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step4.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">5</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step4}
+                                {/* 步驟5 */}
+                            </div>
+                        </div>
+
+                        {/* 分隔線 */}
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step5.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">6</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step5}
+                                {/* 步驟6 */}
+                            </div>
+                        </div>
+
+                        {/* 分隔線 */}
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step6.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">7</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step6}
+                                {/* 步驟7 */}
+                            </div>
+                        </div>
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step7.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">8</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step7}
+                                {/* 步驟8 */}
+                            </div>
+                        </div>
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step8.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">9</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step8}
+                                {/* 步驟9 */}
+                            </div>
+                        </div>
+
+                        <div
+                            className={
+                                eachrecipe.recipes_step9.length === 0
+                                    ? 'showarea2ineach1'
+                                    : 'showarea2ineach'
+                            }
+                        >
+                            <div className="numberineachrecipe">10</div>
+                            <div className="chapter2ineach">
+                                {eachrecipe.recipes_step9}
+                                {/* 步驟10 */}
+                            </div>
+                        </div>
                     </div>
                 </div>
             </div>
@@ -256,10 +570,15 @@ function Eachrecipe() {
                     </button>
                 </Link>
 
+                <button className="hsiehupdate" onClick={sendtoupdate}>
+                    修改食譜
+                    <img src="/images/pen.svg" alt="" className="crudineach" />
+                </button>
+
                 <Link to={`http://localhost:3600/recipe/delete`}>
                     <button
                         className="hsiehdelete"
-                        onClick={(e) => Delectrecipe()}
+                        onClick={(e) => Deleterecipe()}
                     >
                         刪除食譜
                         <img
